+++ pybamm/parameters/parameter_values.py
@@ -48,10 +48,10 @@ class ParameterValues:
     1
     >>> file = "input/parameters/lithium_ion/cells/kokam_Marquis2019/parameters.csv"
     >>> values_path = pybamm.get_parameters_filepath(file)
-    >>> param = pybamm.ParameterValues(values=values_path)
+    >>> param = pybamm.ParameterValues(values_path)
     >>> param["Negative current collector thickness [m]"]
     2.5e-05
-    >>> param = pybamm.ParameterValues(chemistry=pybamm.parameter_sets.Marquis2019)
+    >>> param = pybamm.ParameterValues("Marquis2019")
     >>> param["Reference temperature [K]"]
     298.15
 
@@ -64,25 +64,37 @@ class ParameterValues:
             raise ValueError(
                 "Only one of values and chemistry can be provided. To change parameters"
                 " slightly from a chemistry, first load parameters with the chemistry"
-                " (param = pybamm.ParameterValues(chemistry=...)) and then update with"
+                " (param = pybamm.ParameterValues(...)) and then update with"
                 " param.update({dict of values})."
             )
         if values is None and chemistry is None:
             raise ValueError("values and chemistry cannot both be None")
         # First load chemistry
         if chemistry is not None:
+            warnings.warn(
+                "The 'chemistry' keyword argument has been deprecated and will be "
+                "removed in a future release. Call `ParameterValues` with a "
+                "parameter set dictionary, or the name of a parameter set (string), "
+                "as the single argument, e.g. `ParameterValues('Chen2020')`.",
+                DeprecationWarning,
+            )
             self.update_from_chemistry(chemistry)
         # Then update with values dictionary or file
         if values is not None:
-            # If base_parameters is a filename, load from that filename
-            if isinstance(values, str):
-                file_path = self.find_parameter(values)
-                path = os.path.split(file_path)[0]
-                values = self.read_parameters_csv(file_path)
+            if (isinstance(values, str) and hasattr(pybamm.parameter_sets, values)) or (
+                isinstance(values, dict) and "chemistry" in values
+            ):
+                self.update_from_chemistry(values)
             else:
-                path = ""
-            # Don't check parameter already exists when first creating it
-            self.update(values, check_already_exists=False, path=path)
+                # If base_parameters is a filename, load from that filename
+                if isinstance(values, str):
+                    file_path = self.find_parameter(values)
+                    path = os.path.split(file_path)[0]
+                    values = self.read_parameters_csv(file_path)
+                else:
+                    path = ""
+                # Don't check parameter already exists when first creating it
+                self.update(values, check_already_exists=False, path=path)
 
         # Initialise empty _processed_symbols dict (for caching)
         self._processed_symbols = {}
@@ -108,6 +120,9 @@ class ParameterValues:
     def __repr__(self):
         return pformat(self._dict_items, width=1)
 
+    def __eq__(self, other):
+        return self._dict_items == other._dict_items
+
     def keys(self):
         """Get the keys of the dictionary"""
         return self._dict_items.keys()
@@ -123,7 +138,7 @@ class ParameterValues:
     def copy(self):
         """Returns a copy of the parameter values. Makes sure to copy the internal
         dictionary."""
-        return ParameterValues(values=self._dict_items.copy())
+        return ParameterValues(self._dict_items.copy())
 
     def search(self, key, print_values=True):
         """
@@ -137,6 +152,9 @@ class ParameterValues:
         """
         Load standard set of components from a 'chemistry' dictionary
         """
+        if isinstance(chemistry, str):
+            chemistry = getattr(pybamm.parameter_sets, chemistry)
+
         base_chemistry = chemistry["chemistry"]
 
         # Load each component name
@@ -159,40 +177,16 @@ class ParameterValues:
             component_groups += ["lithium plating"]
 
         if "anode" in chemistry.keys():
-            if "negative electrode" in chemistry.keys():
-                raise KeyError(
-                    "both 'anode' and 'negative electrode' keys provided in the "
-                    "chemistry. The 'anode' notation will be deprecated in the next "
-                    "release so 'negative electrode' should be used instead."
-                )
-            else:
-                chemistry["negative electrode"] = chemistry["anode"]
-                warnings.warn(
-                    "the 'anode' component notation will be deprecated in the next "
-                    "release, as it has now been renamed to 'negative electrode'. "
-                    "Simulation will continue passing the 'anode' component as "
-                    "'negative electrode' (it might overwrite any existing definition "
-                    "of the component).",
-                    DeprecationWarning,
-                )
+            raise KeyError(
+                "The 'anode' notation has been deprecated, "
+                "'negative electrode' should be used instead."
+            )
 
         if "cathode" in chemistry.keys():
-            if "positive electrode" in chemistry.keys():
-                raise KeyError(
-                    "both 'cathode' and 'positive electrode' keys provided in the "
-                    "chemistry. The 'cathode' notation will be deprecated in the next "
-                    "release so 'positive electrode' should be used instead."
-                )
-            else:
-                chemistry["positive electrode"] = chemistry["cathode"]
-                warnings.warn(
-                    "the 'cathode' component notation will be deprecated in the next "
-                    "release, as it has now been renamed to 'positive electrode'. "
-                    "Simulation will continue passing the 'cathode' component as "
-                    "'positive electrode' (it might overwrite any existing definition "
-                    "of the component).",
-                    DeprecationWarning,
-                )
+            raise KeyError(
+                "The 'cathode' notation has been deprecated, "
+                "'positive electrode' should be used instead."
+            )
 
         for component_group in component_groups:
             # Make sure component is provided
@@ -349,24 +343,10 @@ class ParameterValues:
                 "capacity [A.h]', and used to calculate current from C-rate."
             )
         if "Cell capacity [A.h]" in values:
-            if "Nominal cell capacity [A.h]" in values:
-                raise ValueError(
-                    "both 'Cell capacity [A.h]' and 'Nominal cell capacity [A.h]' "
-                    "provided in values. The 'Cell capacity [A.h]' notation will be "
-                    "deprecated in the next release so 'Nominal cell capacity [A.h]' "
-                    "should be used instead."
-                )
-            else:
-                values["Nominal cell capacity [A.h]"] = values["Cell capacity [A.h]"]
-                warnings.warn(
-                    "the 'Cell capacity [A.h]' notation will be "
-                    "deprecated in the next release, as it has now been renamed "
-                    "to 'Nominal cell capacity [A.h]'. Simulation will continue "
-                    "passing the 'Cell capacity [A.h]' as 'Nominal cell "
-                    "capacity [A.h]' (it might overwrite any existing definition "
-                    "of the component)",
-                    DeprecationWarning,
-                )
+            raise ValueError(
+                "The 'Cell capacity [A.h]' parameter has been deprecated, "
+                "'Nominal cell capacity [A.h]' should be used instead."
+            )
         for param in values:
             if "surface area density" in param:
                 raise ValueError(